requirements.txt
flask
pymongo
pyjwt
Werkzeug
flask-cors
dotenv
python-dotenv
gunicorn
eventlet
